Extracted Player model class and tier_names dictionaries to separate files
